DB Initiation
I started to create the database and am adding data to the tables
dynamically through a txt file. Trying to test this out before writing
all the code for the db but having some trouble trying to get the
database to initiate. I think this is bc when running via phone, the
data pack is created on the phone and not on the hdd so going to have to
test via emulator. Also need to write methods for retrieving data.