Database on its way!
See comments in classes (at the top) for more info:
1. Need to make Bar class have similar setup to Contact class
2. Need to Update the DatabaseHelper class by updating the query strings
to match the ones in DatabaseHelperV1 (right now, they are for
contacts-- need to do bars)
3. Need to parse tokens from input file bars.txt into LoadScreenActivity
using a loop and pushing those tokens into a Bar object. Then insert the
bar objects into the Bar Table.
4. Get REST OF BAR INFORMATION AND PUSH DATA INTO FIRST TABLE!
//
5. Figure out how to handle updating the database ONLY when info has
changed (as opposed to always (re)creating the database when the app is
called from the mobile device.
6. Create other tables to meet the apps needs.